Moved settings to its own file, optimised imports
